# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MarketAPI",
        "contact": {
            "email": "jonathan.juarez.valera@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/store": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Create a new user",
                "description": "Create a new user in the database",
                "operationId": "c28b0193df0c53eb5ed5ace1f349a0c3",
                "requestBody": {
                    "description": "Data to create a user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/user_request"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/user_response"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected error, notify to the developer"
                    }
                }
            }
        },
        "/api/users/update": {
            "patch": {
                "tags": [
                    "users"
                ],
                "summary": "Update a user and remove the current token",
                "description": "Make an user update and remove the token of the user",
                "operationId": "0907da039d279521e4dfcfa920ca7ce3",
                "requestBody": {
                    "description": "Data to update in the user",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/user_request"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User update success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                },
                                "example": 1
                            }
                        }
                    },
                    "401": {
                        "description": "The user must be authenticated"
                    }
                }
            }
        },
        "/api/users/verifyEmail/{email}": {
            "get": {
                "tags": [
                    "validators"
                ],
                "description": "Check if the user given exists already in users.",
                "operationId": "87da057d98b1f6698ab78427fc18b032",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "example": "prueba@prueba.com"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Result of evaluation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected error, notify to the developer"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "user_response": {
                "title": "UserDTO",
                "description": "Response of a user in each request related to it",
                "properties": {
                    "id": {
                        "description": "ID of the user",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of user",
                        "type": "string",
                        "example": "John Torres"
                    },
                    "email": {
                        "description": "Email of user",
                        "type": "string",
                        "example": "correo@correo.com"
                    },
                    "profile_picture": {
                        "description": "URL to profile picture",
                        "type": "string",
                        "example": "http://example/example.jpg"
                    },
                    "created_at": {
                        "description": "Date of creation",
                        "type": "string",
                        "example": "2025-07-02 18:38:46"
                    },
                    "updated_at": {
                        "description": "Date of the las update",
                        "type": "string",
                        "example": "2025-07-02 18:38:46"
                    }
                },
                "type": "object"
            },
            "user_request": {
                "title": "UserRequestDTO",
                "description": "DTO to make request related with users",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 6,
                        "example": "John Gonzales Luna",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "minLength": 5,
                        "example": "correo@correo.com",
                        "nullable": true
                    },
                    "password_confirmation": {
                        "example": "root12345",
                        "nullable": true
                    },
                    "password": {
                        "type": "string",
                        "minLength": 8,
                        "example": "root12345",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}